code refactoring : 문제 등록 템플릿의 파일 명을 변경했습니다.
